nucliadb_writer/requirements.txt
starlette-prometheus
uvicorn==0.16.0
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-propagator-b3
fastapi-versioning==0.10.0
fastapi==0.75.0
pydantic==1.9.0
sentry-sdk==1.4.1
pyjwt==2.3.0
mmh3==3.0.0
httpx==0.19.0
types-orjson
types-setuptools
aiohttp==3.7.4
